pyproject.toml
[project]
name = "planie-backend"
version = "0.1.0"
description = "Backend for Plantie app"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.12",
    "ipython>=9.0.2",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",
    "numpy>=1.26.4",
    "pillow>=11.1.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.2",
    "tdqm>=0.0.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "uvicorn>=0.34.0",
]
